+++ app.js
@@ -8,7 +8,8 @@
  */
 
 var path = require('path'),
-    fs = require('fs')
+    fs = require('fs'),
+    RedisPool = require('redis-mpool')
 ;
 
 
@@ -31,7 +32,7 @@ var _ = require('underscore');
 global.environment  = require(__dirname + '/config/environments/' + ENV);
 global.environment.api_hostname = require('os').hostname().split('.')[0];
 
-global.log4js = require('log4js')
+global.log4js = require('log4js');
 log4js_config = {
   appenders: [],
   replaceConsole:true
@@ -65,10 +66,16 @@ if ( global.environment.rollbar ) {
 log4js.configure(log4js_config, { cwd: __dirname });
 global.logger = log4js.getLogger();
 
+var redisOpts = _.extend(global.environment.redis, {
+        name: 'windshaft_cartodb',
+        reportInterval: 5000
+    }),
+    redisPool = new RedisPool(redisOpts);
+
 // Include cartodb_windshaft only _after_ the "global" variable is set
 // See https://github.com/Vizzuality/Windshaft-cartodb/issues/28
-var CartodbWindshaft = require('./lib/cartodb/cartodb_windshaft');
-var serverOptions = require('./lib/cartodb/server_options')();
+var CartodbWindshaft = require('./lib/cartodb/cartodb_windshaft'),
+    serverOptions = require('./lib/cartodb/server_options')(redisPool);
 
 ws = CartodbWindshaft(serverOptions);